+++ src/@types/browser.d.ts
@@ -1,0 +1,7 @@
+import type { SidebarAction } from 'webextension-polyfill'
+
+declare module '@wxt-dev/browser' {
+  namespace Browser {
+    const sidebarAction: SidebarAction.Static
+  }
+}