+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add shortcuts for opening OriginTab in side panel.
+
 ## [1.1.3] - 2026-04-16
 
 ### Added